Add check for max number of allowed characters
According to the crontab man pages, the maximum number of characters in
the command field is 998
https://manpages.debian.org/jessie/cron/crontab.5.en.html. This PR adds
a check to catch lines that are longer than 998. It updates the test
file as well. This change was requested in this issue
https://github.com/lyda/chkcrontab/issues/22 and I have run into issues
with crontab failing because of lines being too long, so having the
linter catch this issue will be a big help.