Fix user crontab regular expression
The existing code causes a LINE_ERROR for cronjobs that consist of a single
command without any parameters. This is due to the fact that the regular
expression expects a username string and a command after that, and the conditional
is_user_crontab code concats these two strings.

Instead of doing that, this commit introduces a separate regular expression
for user crontab lines.